requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
langdetect==1.0.9
pyTelegramBotAPI==4.6.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot-calendar==1.0.5
requests==2.28.1
six==1.16.0
urllib3==1.26.11
